Latest version
commit 773e41320bd610f4ad2596038d56ca2ec451766c
Author: Ben Swanson <benswanson@rakocontrols.com>
Date:   Fri Oct 20 11:26:33 2023 +0100

    uint32 correction

commit 28a334f84ec20058922d5816c4a8266f7990adab
Author: Ben Swanson <benswanson@rakocontrols.com>
Date:   Fri Oct 20 10:03:12 2023 +0100

    Initial mdns state value

commit 0c6c7c90677bd2bacbace60c0d83e61f057c0291
Author: Ben Swanson <benswanson@rakocontrols.com>
Date:   Fri Oct 20 10:02:57 2023 +0100

    Characteristic adjustments

commit 730c099ecd20fd08373525e5142d7c28d628e750
Author: Ben Swanson <benswanson@rakocontrols.com>
Date:   Tue Oct 17 17:19:43 2023 +0100

    Update stateIndex

commit f89a924d130399a127951f26cfed722ebfdde86a
Author: Ben Swanson <benswanson@rakocontrols.com>
Date:   Tue Oct 17 17:09:23 2023 +0100

    Root setStateIndex check started

commit 1f69e070df6448a86b813a8af09ae2f8ea8cb833
Author: Ben Swanson <benswanson@rakocontrols.com>
Date:   Tue Oct 17 15:35:37 2023 +0100

    Make registry public and implement setStateIndex

commit 37d87c46d615b77e6fb246a933f073d284987bb2
Author: Ben Swanson <benswanson@rakocontrols.com>
Date:   Tue Oct 17 15:16:37 2023 +0100

    Update to master

+++ src/main/java/io/github/hapjava/server/impl/HomekitRegistry.java
@@ -21,7 +21,7 @@ public class HomekitRegistry {
 
   private final String label;
   private final SubscriptionManager subscriptions;
-  private final Map<Integer, HomekitAccessory> accessories;
+  private final Map<Long, HomekitAccessory> accessories;
   private final Map<HomekitAccessory, Map<Integer, Service>> services = new HashMap<>();
   private final Map<HomekitAccessory, Map<Integer, Characteristic>> characteristics =
       new HashMap<>();
